Refactor yahrzeit download into dedicated module
Split yahrzeitDownload out of yahrzeit.js into yahrzeitDownload.js so that
app-download.js no longer pulls in www-only code. Shared event-generation
logic (makeYahrzeitEvents, getDateRange, makeEditMemo, etc.) moves to a new
makeYahrzeitEvents.js consumed by both yahrzeit.js (www) and
yahrzeitDownload.js (download), yielding a clean import tree with no cycles.
