+++ src/yahrzeit.js
@@ -1,8 +1,5 @@
-import {Event, HDate, HebrewCalendar, Locale,
-  Location,
-  flags, gematriya, months} from '@hebcal/core';
-import {IcalEvent, icalEventsToString} from '@hebcal/icalendar';
-import {eventsToCsv, eventsToClassicApi} from '@hebcal/rest-api';
+import {HDate, months} from '@hebcal/core';
+import {eventsToClassicApi} from '@hebcal/rest-api';
 import dayjs from 'dayjs';
 import {basename} from 'node:path';
 import {empty} from './empty.js';
@@ -11,24 +8,21 @@ import {pkg} from './pkg.js';
 import {cleanQuery} from './cleanQuery.js';
 import {
   hebcalFormDefaults,
-  makeIcalOpts,
   processCookieAndQuery,
 } from './urlArgs.js';
 import {doesCookieNeedRefresh, setHebcalCookie} from './cookie.js';
 import {getIpAddress} from './getIpAddress.js';
-import {lightCandlesWhen} from './common.js';
 import {ulid} from 'ulid';
 import {getMaxYahrzeitId, isNumKey, summarizeAnniversaryTypes,
   getAnniversaryTypes,
-  YAHRZEIT, BIRTHDAY,
   DEFAULT_YEARS,
   getNumYears,
   compactJsonToSave,
   getCalendarNames, makeCalendarTitle,
   getYahrzeitDetailsFromDb, getYahrzeitDetailForId} from './yahrzeitCommon.js';
+import {makeYahrzeitEvents} from './makeYahrzeitEvents.js';
 import {makeLogInfo} from './logger.js';
 import {isDeepStrictEqual} from 'node:util';
-import {murmur128HexSync, murmur32HexSync} from 'murmurhash3';
 
 const urlPrefix = process.env.NODE_ENV == 'production' ? 'https://download.hebcal.com' : 'http://127.0.0.1:8081';
 const TIMEOUT = 7000;
@@ -383,418 +377,4 @@ function removeEmptyArgs(q) {
       }
     }
   }
-}
-
-/**
- * @param {Koa.ParameterizedContext<Koa.DefaultState, Koa.DefaultContext>} ctx
- */
-async function getDetailsFromDb(ctx) {
-  const id = ctx.request.path.substring(4, 30);
-  const obj = await getYahrzeitDetailsFromDb(ctx, id);
-  ctx.state.relcalid = id;
-  return obj;
-}
-
-const maxEventsIcsSub = 1200;
-
-/**
- * @param {Koa.ParameterizedContext<Koa.DefaultState, Koa.DefaultContext>} ctx
- */
-export async function yahrzeitDownload(ctx) {
-  const rpath = ctx.request.path;
-  const details = rpath.startsWith('/v3') ? await getDetailsFromDb(ctx) : {};
-  const query = {...details, ...ctx.request.query};
-  // Fix for legacy duplicated key/value pairs
-  for (const [key, value] of Object.entries(query)) {
-    if (Array.isArray(value) && value.length === 2) {
-      query[key] = value[1];
-    }
-  }
-  const vv = query.v;
-  if (vv?.[0] !== 'y') {
-    return;
-  }
-  if (ctx.state.ulid) {
-    query.ulid = ctx.state.ulid;
-  }
-  const extension = rpath.substring(rpath.length - 4);
-  const ics = extension === '.ics';
-  const csv = extension === '.csv';
-  const {today, startYear, endYear} = getDateRange(query);
-  const sunday = today.onOrBefore(0).abs();
-  const attrs = {startYear, endYear, extension};
-  const isAttachment = query.dl == '1';
-  if (ics) {
-    attrs.icalv = IcalEvent.version();
-    if (!isAttachment) {
-      attrs.sunday = sunday;
-    }
-  }
-  ctx.response.etag = makeETag(ctx, query, attrs);
-  ctx.status = 200;
-  if (ctx.fresh) {
-    ctx.status = 304;
-    return;
-  }
-
-  let reminder = query.yrem !== '0' && query.yrem !== 'off';
-  if (csv) {
-    reminder = false;
-  }
-  const maxId = getMaxYahrzeitId(query);
-  const events0 = await makeYahrzeitEvents(maxId, query, reminder);
-  const startAbs = sunday - 12 * 7; // 12 weeks ago
-  const events = !ics || isAttachment ? events0 : events0.filter((ev) => ev.getDate().abs() >= startAbs);
-
-  if (isAttachment) {
-    ctx.response.attachment(basename(rpath));
-  }
-  if (ics) {
-    ctx.response.type = 'text/calendar; charset=utf-8';
-    const caldesc = makeCalendarTitle(query, Number.POSITIVE_INFINITY);
-    const title = makeCalendarTitle(query, 64);
-    const opts = {
-      yahrzeit: true,
-      emoji: true,
-      title: title,
-      caldesc: caldesc === title ?
-        'Personal anniversaries from Hebcal.com' :
-        'Hebcal ' + caldesc,
-      relcalid: ctx.state.relcalid ? `hebcal-${ctx.state.relcalid}` : null,
-      publishedTTL: 'P1D',
-      sequence: +(query.seq) || 1,
-      locale: 'en',
-    };
-    const icalOpt = makeIcalOpts(opts, query);
-    icalOpt.dtstamp = IcalEvent.makeDtstamp(new Date());
-    // Hack for Google Calendar which doesn't understand iCalendar floating times
-    const typesSet = getAnniversaryTypes(query);
-    const doLocation = reminder && typesSet.has(YAHRZEIT);
-    if (doLocation) {
-      if (query.i === 'on') {
-        icalOpt.location = Location.lookup('Jerusalem');
-      } else if (ctx.get('user-agent') === 'Google-Calendar-Importer') {
-        icalOpt.location = makeLocation(query);
-      }
-    }
-    const zeroEvents = events.length === 0;
-    const events1 = events.length > maxEventsIcsSub && !isAttachment ? events.slice(0, maxEventsIcsSub) : events;
-    const events2 = zeroEvents ? makeDummyEvent(ctx) : events1;
-    if (zeroEvents) {
-      icalOpt.publishedTTL = false;
-    }
-    const icals = events2.map((ev) => new IcalEvent(ev, icalOpt));
-    if (doLocation) {
-      for (const icalEv of icals) {
-        icalEv.locationName = undefined;
-      }
-    }
-    ctx.append('Vary', 'User-Agent');
-    ctx.body = await icalEventsToString(icals, icalOpt);
-  } else if (csv) {
-    const euro = Boolean(query.euro);
-    const csv = eventsToCsv(events, {euro});
-    ctx.response.type = 'text/x-csv; charset=utf-8';
-    ctx.body = '\uFEFF' + csv;
-  } else {
-    ctx.status = 404;
-    ctx.remove('ETag');
-    ctx.remove('Cache-Control');
-    ctx.response.type = 'text/plain';
-    ctx.body = 'Invalid download format: ' + extension + '\n';
-  }
-}
-
-function makeLocation(query) {
-  const tzo = Number.parseInt(query.tzo, 10);
-  if (Number.isNaN(tzo)) {
-    return Location.lookup('New York');
-  }
-  const tz = tzo / -60;
-  let tzid;
-  if (tz <= -4 && tz >= -10) {
-    tzid = Location.legacyTzToTzid(tz, 'usa');
-  } else if (tz <= 2 && tz >= -2) {
-    tzid = Location.legacyTzToTzid(tz, 'eu');
-  } else {
-    tzid = Location.legacyTzToTzid(tz, 'none');
-  }
-  return new Location(0, 0, false, tzid, '');
-}
-
-function makeEditMemo(calendarId) {
-  return 'To edit this Hebcal Yahrzeit + Anniversary Calendar, visit https://www.hebcal.com/yahrzeit/edit/' +
-    calendarId;
-}
-
-function makeDummyEvent(ctx) {
-  const dt = new Date();
-  const ev = new Event(new HDate(dt), 'Calendar contains no events', flags.USER_EVENT);
-  const url = ctx.request.url;
-  const id = ctx.state.relcalid || murmur128HexSync(url);
-  const isoDateStr = IcalEvent.formatYYYYMMDD(dt);
-  ev.uid = `yahrzeit-${isoDateStr}-${id}-dummy`;
-  ev.alarm = false;
-  if (ctx.state.relcalid) {
-    const id = ctx.state.relcalid;
-    ev.memo = makeEditMemo(id);
-  } else {
-    ev.memo = 'To create a new Hebcal Yahrzeit + Anniversary Calendar, visit https://www.hebcal.com/yahrzeit';
-  }
-  ctx.set('Cache-Control', 'max-age=86400');
-  ctx.remove('ETag');
-  return [ev];
-}
-
-/**
- * @return {number}
- */
-function getDefaultStartYear(today) {
-  const hmonth = today.getMonth();
-  const hyear = today.getFullYear();
-  const isFirst3months = hmonth >= months.TISHREI && hmonth <= months.TEVET;
-  return isFirst3months ? hyear - 1 : hyear;
-}
-
-function getDateRange(query) {
-  const today = new HDate();
-  const years = getNumYears(query.years);
-  const startYear = Number.parseInt(query.start, 10) || getDefaultStartYear(today);
-  const endYear = Number.parseInt(query.end, 10) || (startYear + years - 1);
-  return {today, startYear, endYear, years};
-}
-
-/**
- * @param {number} maxId
- * @param {any} query
- * @param {boolean} reminder
- * @return {Promise<Event[]>}
- */
-async function makeYahrzeitEvents(maxId, query, reminder) {
-  const {startYear, endYear, years} = getDateRange(query);
-  let events = [];
-  for (let id = 1; id <= maxId; id++) {
-    const events0 = await getEventsForId(query, id, startYear, years);
-    events = events.concat(events0);
-    if (reminder) {
-      const reminders = makeReminderEvents(events0, id);
-      events = events.concat(reminders);
-    }
-  }
-  const yizkor = query.yizkor;
-  if (yizkor === 'on' || yizkor === '1') {
-    const holidays = makeYizkorEvents(startYear, endYear, query.i === 'on');
-    for (const ev of holidays) {
-      const d = dayjs(ev.greg());
-      const hash = murmur32HexSync(ev.getDesc());
-      ev.uid = 'yizkor-' + d.format('YYYYMMDD') + '-' + hash;
-      if (query.ulid) {
-        ev.memo = makeEditMemo(query.ulid);
-      }
-    }
-    events = events.concat(holidays);
-  }
-  events.sort((a, b) => a.getDate().abs() - b.getDate().abs());
-  return events;
-}
-
-function getAlarmTime(hd) {
-  switch (hd.getDay()) {
-    case 6: return '20:00';
-    case 5: return '14:30';
-    default: return '16:30';
-  }
-}
-
-function makeReminderEvents(events, id) {
-  const yahrzeitEvents = events.filter((ev) => ev.type === YAHRZEIT);
-  return yahrzeitEvents.map((ev) => {
-    const hd = ev.getDate().prev();
-    const dt = hd.greg();
-    const uid = 'reminder-' + dayjs(dt).format('YYYYMMDD') + '-' + ev.hash + '-' + id;
-    const name = ev.name;
-    const subj = hebrewRe.test(name) ?
-      `${name} ${YAHRZEIT_HE} תזכורת` :
-      `${name} Yahrzeit reminder`;
-    return new Event(hd, subj, flags.USER_EVENT, {
-      eventTime: dt,
-      eventTimeStr: getAlarmTime(hd),
-      memo: ev.memo,
-      emoji: ev.emoji,
-      alarm: 'P0DT0H0M0S',
-      uid,
-      category: 'Personal',
-    });
-  });
-}
-
-/**
- * @param {any} query
- * @param {number} id
- * @param {number} startYear
- * @param {number} numYears
- * @return {Promise<Event[]>}
- */
-async function getEventsForId(query, id, startYear, numYears) {
-  const events = [];
-  const info = getYahrzeitDetailForId(query, id);
-  if (info === null) {
-    return events;
-  }
-  const calendarId = query.ulid;
-  const includeUrl = Boolean(calendarId && query.dl !== '1');
-  const appendHebDate = (query.hebdate === 'on' || query.hebdate === '1');
-  for (let hyear = startYear; events.length < numYears; hyear++) {
-    const ev = await makeYahrzeitEvent(id, info, hyear, appendHebDate, calendarId, includeUrl);
-    if (ev) {
-      events.push(ev);
-    }
-  }
-  return events;
-}
-
-const hebrewRe = /[\u05d0-\u05ea]/;
-
-const YAHRZEIT_HE = 'יארצייט';
-
-const en2he = {
-  Yahrzeit: YAHRZEIT_HE,
-  Birthday: 'יום הולדת',
-  Anniversary: 'יום נישואין',
-};
-
-function hebdateNoYear(hd, isHebrewName) {
-  if (isHebrewName) {
-    const dd = hd.getDate();
-    const mm = Locale.gettext(hd.getMonthName(), 'he-x-NoNikud');
-    return gematriya(dd) + ' ' + mm;
-  } else {
-    return hd.render('en', false).replaceAll('\'', '’');
-  }
-}
-
-function makeYahrzeitSubject(info, hd, yearNumber, appendHebDate) {
-  const name = info.name;
-  let subj = name;
-  const isHebrewName = hebrewRe.test(name) && !/[A-Za-z]/.test(name);
-  const type = info.type;
-  if (type !== 'Other') {
-    const isYahrzeit = type === YAHRZEIT;
-    if (isHebrewName) {
-      const prefix = en2he[type];
-      subj = isYahrzeit ?
-        `${prefix} ה-${yearNumber} של ${name}` :
-        `${prefix} ${yearNumber} ל${name}`;
-    } else {
-      const nth = Locale.ordinal(yearNumber, 'en');
-      const typeStr = isYahrzeit ? type : `Hebrew ${type}`;
-      subj = `${name}’s ${nth} ${typeStr}`;
-    }
-  }
-  if (appendHebDate) {
-    const hebdate = hebdateNoYear(hd, isHebrewName);
-    subj += ' (' + hebdate + ')';
-  }
-  return subj;
-}
-
-/**
- * @param {number} id
- * @param {any} info
- * @param {number} hyear
- * @param {boolean} appendHebDate
- * @param {string} calendarId
- * @param {boolean} includeUrl
- * @return {Promise<Event>}
- */
-async function makeYahrzeitEvent(id, info, hyear, appendHebDate, calendarId, includeUrl) {
-  const type = info.type;
-  const isYahrzeit = type === YAHRZEIT;
-  const isBirthday = type === BIRTHDAY;
-  const origDt = info.day.toDate();
-  const hd = isYahrzeit ?
-    HebrewCalendar.getYahrzeit(hyear, origDt) :
-    HebrewCalendar.getBirthdayOrAnniversary(hyear, origDt);
-  if (!hd) {
-    return null;
-  }
-  const typeStr = isYahrzeit ? type : `Hebrew ${type}`;
-  const hebdate = hd.render('en').replaceAll('\'', '’');
-  const origHd = new HDate(origDt);
-  const origHyear = origHd.getFullYear();
-  const yearNumber = hyear - origHyear;
-  const nth = Locale.ordinal(yearNumber, 'en');
-  const name = info.name;
-  const subj = makeYahrzeitSubject(info, hd, yearNumber, appendHebDate);
-  const ev = new Event(hd, subj, flags.USER_EVENT);
-  if (isYahrzeit) {
-    ev.emoji = '🕯️';
-  } else if (isBirthday) {
-    ev.emoji = '🎂✡️';
-  }
-  const observed = dayjs(hd.greg());
-  ev.memo = makeMemo(id, info, observed, nth, typeStr, hebdate, includeUrl, calendarId);
-  const hash = calendarId || murmur32HexSync(name);
-  ev.uid = type.toLowerCase() + '-' + hyear + '-' + hash + '-' + id;
-  ev.name = name;
-  ev.type = type;
-  ev.anniversary = yearNumber;
-  if (isYahrzeit) {
-    ev.alarm = false;
-    ev.hash = hash;
-  }
-  return ev;
-}
-
-function makeMemo(id, info, observed, nth, typeStr, hebdate, includeUrl, calendarId) {
-  const type = info.type;
-  const isYahrzeit = type === YAHRZEIT;
-  const isBirthday = type === BIRTHDAY;
-  const isOther = (type === 'Other');
-  const name = info.name;
-  const nameAndType = isOther ? name : `${name}’s ${typeStr}`;
-  const erev = observed.subtract(1, 'day');
-  const verb = isYahrzeit ? 'remembering' : 'honoring';
-  const prefix = isOther ? name : `Hebcal joins you in ${verb} ${name}, whose ${nth} ${typeStr}`;
-  let memo = `${prefix} occurs on ` +
-    `${observed.format('dddd, MMMM D')}, corresponding to the ${hebdate}.\\n\\n` +
-    `${nameAndType} begins at sundown on ${erev.format('dddd, MMMM D')} and continues until ` +
-    `sundown on the day of observance.`;
-  if (isYahrzeit) {
-    const dow = erev.day();
-    const when = lightCandlesWhen(dow);
-    memo += ` It is customary to light a memorial candle ${when} as the Yahrzeit begins.\\n\\n` +
-      'May your loved one’s soul be bound up in the bond of eternal life and may their memory ' +
-      'serve as a continued source of inspiration and comfort to you.';
-  } else if (isBirthday) {
-    memo += '\\n\\nMazel Tov!';
-  }
-  if (includeUrl) {
-    memo += `\\n\\nhttps://www.hebcal.com/yahrzeit/edit/${calendarId}#row${id}`;
-  }
-  return memo;
-}
-
-/**
- * @param {number} startYear
- * @param {number} endYear
- * @param {boolean} il
- * @return {Event[]}
- */
-function makeYizkorEvents(startYear, endYear, il) {
-  const holidays = [];
-  const attrs = {emoji: '🕯️'};
-  const pesachDay = il ? 21 : 22;
-  const pesachDesc = il ? 'Yizkor (Pesach VII)' : 'Yizkor (Pesach VIII)';
-  const shavuotDay = il ? 6 : 7;
-  const shavuotDesc = il ? 'Yizkor (Shavuot)' : 'Yizkor (Shavuot II)';
-  for (let hyear = startYear; hyear <= endYear; hyear++) {
-    holidays.push(
-        new Event(new HDate(pesachDay, months.NISAN, hyear), pesachDesc, flags.USER_EVENT, attrs),
-        new Event(new HDate(shavuotDay, months.SIVAN, hyear), shavuotDesc, flags.USER_EVENT, attrs),
-        new Event(new HDate(10, months.TISHREI, hyear), 'Yizkor (Yom Kippur)', flags.USER_EVENT, attrs),
-        new Event(new HDate(22, months.TISHREI, hyear), 'Yizkor (Shmini Atzeret)', flags.USER_EVENT, attrs),
-    );
-  }
-  return holidays;
 }
